fix(extension-usage): extract usage utility types
