+++ packages/extension-usage/src/index.ts
@@ -1,39 +1,15 @@
 import { resolve } from 'path'
-import { Context, h, Logger, Schema, Time } from 'koishi'
+import { Context, h, Logger, Time } from 'koishi'
 import type { Session } from 'koishi'
 import { DataService } from '@koishijs/plugin-console'
-import type { UsageMetadata } from '@langchain/core/messages'
-import type { ModelUsageCallType } from 'koishi-plugin-chatluna/llm-core/platform/usage'
+import { ChatLunaUsage, summary } from './utils'
 import type {} from 'koishi-plugin-puppeteer'
 import { renderTokenTrend } from './renderer'
 import { createTokenReport, formatTokenReport } from './tokens'
-import type { TokenRange, TokenTheme } from './tokens'
 
 const logger = new Logger('chatluna-usage')
 
-function summary(
-    key: string,
-    label = key,
-    platform?: string
-): ChatLunaUsage.Summary {
-    return {
-        key,
-        label,
-        platform,
-        calls: 0,
-        successfulCalls: 0,
-        failedCalls: 0,
-        inputTokens: 0,
-        outputTokens: 0,
-        totalTokens: 0,
-        estimatedTokens: 0,
-        cachedTokens: 0,
-        reasoningTokens: 0,
-        successRate: 0
-    }
-}
-
-class ChatLunaUsage extends DataService<ChatLunaUsage.Payload> {
+class ChatLunaUsageService extends DataService<ChatLunaUsage.Payload> {
     constructor(
         ctx: Context,
         public config: ChatLunaUsage.Config
@@ -262,13 +238,18 @@ class ChatLunaUsage extends DataService<ChatLunaUsage.Payload> {
         options: ChatLunaUsage.TokenCommandOptions,
         args: string[]
     ) {
-        let range: TokenRange = options.all
-            ? 'all'
-            : options.month
-              ? 'month'
-              : options.week
-                ? 'week'
-                : 'day'
+        let range: ChatLunaUsage.TokenRange
+
+        if (options.all) {
+            range = 'all'
+        } else if (options.month) {
+            range = 'month'
+        } else if (options.week) {
+            range = 'week'
+        } else {
+            range = 'day'
+        }
+
         let plugin = Boolean(options.plugin)
 
         for (const arg of args) {
@@ -282,7 +263,7 @@ class ChatLunaUsage extends DataService<ChatLunaUsage.Payload> {
                 month: 'month',
                 a: 'all',
                 all: 'all'
-            }[keyword] as TokenRange
+            }[keyword] as ChatLunaUsage.TokenRange
             if (value) {
                 range = value
                 continue
@@ -323,7 +304,10 @@ class ChatLunaUsage extends DataService<ChatLunaUsage.Payload> {
         }
     }
 
-    private async tokenReport(range: TokenRange, withPlugins = false) {
+    private async tokenReport(
+        range: ChatLunaUsage.TokenRange,
+        withPlugins = false
+    ) {
         const end = new Date()
         const start = new Date(
             +end -
@@ -354,10 +338,7 @@ class ChatLunaUsage extends DataService<ChatLunaUsage.Payload> {
         if (query.success != null) where.success = query.success
         if (query.estimated != null) where.estimated = query.estimated
 
-        const rows = (await this.ctx.database.get(
-            'chatluna_usage',
-            where
-        )) as ChatLunaUsage.Record[]
+        const rows = await this.ctx.database.get('chatluna_usage', where)
 
         if (
             !query.chatPlatform &&
@@ -516,209 +497,11 @@ class ChatLunaUsage extends DataService<ChatLunaUsage.Payload> {
     }
 }
 
-// eslint-disable-next-line @typescript-eslint/no-namespace
-namespace ChatLunaUsage {
-    export interface Record {
-        id?: number
-        source: string
-        callType: ModelUsageCallType
-        platform: string
-        chatPlatform?: string | null
-        model: string
-        usageMetadata: UsageMetadata
-        estimated: boolean
-        success: boolean
-        createdAt: Date
-        conversationId?: string | null
-        requestId?: string | null
-        userId?: string | null
-        guildId?: string | null
-    }
-
-    export interface ListRow extends Record {
-        inputTokens: number
-        outputTokens: number
-        totalTokens: number
-        estimated: boolean
-        cachedTokens: number
-        reasoningTokens: number
-    }
-
-    export type Period = 'day' | 'month' | 'year'
-    export type GroupBy =
-        | 'source'
-        | 'model'
-        | 'guild'
-        | 'platform'
-        | 'chatPlatform'
-        | 'callType'
-    export type SortBy =
-        | 'calls'
-        | 'successfulCalls'
-        | 'failedCalls'
-        | 'inputTokens'
-        | 'outputTokens'
-        | 'totalTokens'
-        | 'estimatedTokens'
-        | 'cachedTokens'
-        | 'reasoningTokens'
-        | 'successRate'
-    export type ListSortBy =
-        | 'createdAt'
-        | 'inputTokens'
-        | 'outputTokens'
-        | 'totalTokens'
-        | 'cachedTokens'
-        | 'reasoningTokens'
-
-    export interface Query {
-        period?: Period
-        start?: string | Date
-        end?: string | Date
-        groupBy?: GroupBy
-        sortBy?: SortBy
-        desc?: boolean
-        page?: number
-        pageSize?: number
-        listSortBy?: ListSortBy
-        listDesc?: boolean
-        source?: string
-        model?: string
-        platform?: string
-        chatPlatform?: string
-        callType?: ModelUsageCallType
-        guildId?: string
-        userId?: string
-        success?: boolean
-        estimated?: boolean
-        keyword?: string
-    }
-
-    export interface Summary {
-        key: string
-        label: string
-        platform?: string
-        calls: number
-        successfulCalls: number
-        failedCalls: number
-        inputTokens: number
-        outputTokens: number
-        totalTokens: number
-        estimatedTokens: number
-        cachedTokens: number
-        reasoningTokens: number
-        successRate: number
-        lastSeen?: Date
-    }
-
-    export interface Timeline {
-        date: string
-        calls: number
-        inputTokens: number
-        outputTokens: number
-        totalTokens: number
-        cachedTokens: number
-        reasoningTokens: number
-    }
-
-    export interface ModelTimeline {
-        model: string
-        points: {
-            date: string
-            calls: number
-        }[]
-    }
-
-    export interface List {
-        total: number
-        page: number
-        pageSize: number
-        rows: ListRow[]
-    }
-
-    export interface Payload {
-        query: Required<
-            Pick<
-                Query,
-                | 'period'
-                | 'groupBy'
-                | 'sortBy'
-                | 'desc'
-                | 'page'
-                | 'pageSize'
-                | 'listSortBy'
-                | 'listDesc'
-            >
-        > & {
-            start: Date
-            end: Date
-        } & Query
-        totals: Summary
-        groups: Summary[]
-        models: Summary[]
-        sources: Summary[]
-        timeline: Timeline[]
-        modelTimeline: ModelTimeline[]
-        list: List
-    }
-
-    export interface Config {
-        recentDays: number
-        pageSize: number
-        webui: boolean
-        tokensTheme: TokenTheme
-    }
-
-    export interface TokenCommandOptions {
-        day?: boolean
-        week?: boolean
-        month?: boolean
-        all?: boolean
-        plugin?: boolean
-    }
-
-    export interface ActionResult {
-        success: boolean
-    }
-
-    export const Config: Schema<Config> = Schema.object({
-        recentDays: Schema.natural()
-            .description('默认统计最近几天的数据。')
-            .default(30),
-        pageSize: Schema.natural()
-            .description('调用明细分页大小。')
-            .default(50),
-        webui: Schema.boolean()
-            .description('启用 Web UI 控制台用量面板。')
-            .default(true),
-        tokensTheme: Schema.union([
-            Schema.const('auto').description('自动'),
-            Schema.const('light').description('浅色模式'),
-            Schema.const('dark').description('深色模式')
-        ])
-            .description('tokens命令渲染出的图表颜色主题')
-            .default('auto')
-            .role('select')
-    })
-
-    export const inject = ['chatluna', 'database']
-}
-
-export default ChatLunaUsage
-export { ChatLunaUsage }
-
-export async function queryUsage(ctx: Context, source?: string) {
-    const result = await ctx.chatluna_usage.query({ groupBy: 'source' })
-    if (!source) return result.groups
-    return result.groups.filter((row) => row.key === source)
-}
-
-export async function cleanupUsage(ctx: Context, before?: Date) {
-    await ctx.chatluna_usage.cleanup(before)
-}
+export default ChatLunaUsageService
+export { ChatLunaUsageService as ChatLunaUsage }
 
 export function apply(ctx: Context, config: ChatLunaUsage.Config) {
-    ctx.plugin(ChatLunaUsage, config)
+    ctx.plugin(ChatLunaUsageService, config)
 }
 
 export const Config = ChatLunaUsage.Config
@@ -732,7 +515,7 @@ export const name = 'chatluna-usage'
 
 declare module 'koishi' {
     interface Context {
-        chatluna_usage: ChatLunaUsage
+        chatluna_usage: ChatLunaUsageService
     }
 
     interface Tables {
@@ -744,7 +527,7 @@ declare module '@koishijs/plugin-console' {
     // eslint-disable-next-line @typescript-eslint/no-namespace
     namespace Console {
         interface Services {
-            chatluna_usage: ChatLunaUsage
+            chatluna_usage: ChatLunaUsageService
         }
     }
